add logic create createLead

+++ src/main/java/com/muf/common/exception/customleadflow/LeadNotFoundException.java
@@ -3,7 +3,7 @@ package com.muf.common.exception.customleadflow;
 import com.muf.base.exception.BusinessException;
 
 public class LeadNotFoundException extends BusinessException {
-    public LeadNotFoundException(Long id) {
+    public LeadNotFoundException(Integer id) {
         super("Lead not found with id: " + id);
     }
 }
