+++ src/main/java/com/muf/common/exception/customleadflow/UserNotFoundException.java
@@ -3,7 +3,7 @@ package com.muf.common.exception.customleadflow;
 import com.muf.base.exception.BusinessException;
 
 public class UserNotFoundException extends BusinessException {
-    public UserNotFoundException(Long id) {
+    public UserNotFoundException(Integer id) {
         super("User not found with id: " + id);
     }
 }
