+++ src/main/java/com/muf/modules/master/lead/dto/LeadResponse.java
@@ -1,6 +1,8 @@
 package com.muf.modules.master.lead.dto;
 
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.time.LocalDateTime;
 
@@ -13,7 +15,7 @@ public class LeadResponse {
     private String sourceCode;
     private String sourceName;
     private String status;
-    private Long ownerId;
+    private Integer ownerId;
     private String ownerName;
     private LocalDateTime createdAt;
 }
